move cairo out, turshi to each folder of kimchi, verification with cs
