+++ cairo/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "cairo"
+version = "0.1.0"
+edition = "2021"
+license = "Apache-2.0"
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
+ark-ff = { version = "0.3.0", features = [ "parallel", "asm" ] }
+ark-ec = { version = "0.3.0", features = [ "parallel" ] }
+hex = "0.4"
+
+mina-curves = { path = "../curves" }
+o1-utils = { path = "../utils" }
