+++ cairo/src/helper.rs
