+++ cairo/src/word.rs
@@ -4,8 +4,8 @@
 //! Our Pallas curves have 255 bits, so Cairo native instructions will fit.
 //! This means that our Cairo implementation can admit a larger domain for immediate values than theirs.
 
-use crate::circuits::turshi::cairo::flags::*;
-use crate::circuits::turshi::cairo::helper::CairoFieldHelpers;
+use crate::flags::*;
+use crate::helper::CairoFieldHelpers;
 use ark_ff::Field;
 use o1_utils::field_helpers::FieldHelpers;
 
@@ -255,8 +255,8 @@ impl<F: Field> FlagSets<F> for CairoWord<F> {
 
 #[cfg(test)]
 mod tests {
-    use crate::circuits::turshi::cairo::flags::*;
-    use crate::circuits::turshi::cairo::word::{FlagBits, FlagSets, Offsets};
+    use crate::flags::*;
+    use crate::word::{FlagBits, FlagSets, Offsets};
     use ark_ff::{One, Zero};
     use mina_curves::pasta::fp::Fp as F;
