+++ kimchi/src/circuits/gates/mod.rs
@@ -1,3 +1,4 @@
+pub mod cairo;
 pub mod endosclmul;
 pub mod poseidon;
 pub mod varbasemul;
