+++ kimchi/src/circuits/mod.rs
@@ -10,5 +10,4 @@ pub mod gates;
 pub mod polynomial;
 pub mod polynomials;
 pub mod scalars;
-pub mod turshi;
 pub mod wires;
