--- kimchi/src/circuits/turshi/mod.rs
@@ -1,6 +1,0 @@
-//! This module contains the code that expresses Cairo gates in Kimchi format
-#![warn(missing_docs)]
-pub mod cairo;
-pub mod cairogate;
-pub mod constraint;
-pub mod witness;
